Kaggle Ultrasound Nerve Segmentation competition
The task in this competition is to segment a collection of nerves called
the Brachial Plexus (BP) in ultrasound images. You are provided with a
large training set of images where the nerve has been manually annotated
by humans. Annotators were trained by experts and instructed to annotate
images where they felt confident about the existence of the BP landmark.

The model was build based on the U-Net: Convolutional Networks for
Biomedical Image Segmentation
[U-Net](https://lmb.informatik.uni-freiburg.de/people/ronneber/u-net/)

/train/ contains the training set images, named according to
subject_imageNum.tif. Every image with the same subject number comes
from the same person. This folder also includes binary mask images
showing the BP segmentations.

/test/ contains the test set images, named according to imageNum.tif.
You must predict the BP segmentation for these images and are not
provided a subject number. There is no overlap between the subjects in
the training and test sets.

**train_masks.csv** gives the training image masks in run-length encoded
format. This is provided as a convenience to demonstrate how to turn
image masks into encoded text values for submission.

sample_submission.csv shows the correct submission file format.